refactor: split `models.py` into modules
